Added ability to recognize and respect #AUTO_UPDATE line in unmenu files
#AUTO_UPDATE = NO
The above will indicate the file will NOT be automatically updated, but a note will be output in the
resulting release printout to the screen.  The new file will be in the /tmp/unmenu_tmp directory.

#AUTO_UPDATE = YES
The above will indicate the file will be automatically updated, but a note will be output in the
resulting release printout to the screen.  The old file will be in the unmenu_old_YYYY-MM-DD_HHMI  directory.